Relax memory orderings
For the BSEC we need to use Acquire and Release to ensure the
manipulations of the unsafe bsec library are not order outside of the
the area holding the BSEC_IN_USE lock. However, no AcqRel is required
because we need to establish no memory access get ordered after the
write of the lock.

For the FakeClock there is no critical section, purely incrementing the
counter in an atomic way is sufficient. Thus, relaxed ordering is fine
here.